modify: change java to kt (AppRepository.kt)
