+++ src/main/kotlin/info/logbat/domain/log/application/LogService.java
@@ -3,7 +3,7 @@ package info.logbat.domain.log.application;
 import info.logbat.common.event.EventProducer;
 import info.logbat.domain.log.domain.Log;
 import info.logbat.domain.log.presentation.payload.request.CreateLogRequest;
-import info.logbat.domain.project.application.AppService;
+import info.logbat.domain.project.application.AppService_ASIS;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,7 +17,7 @@ import java.util.List;
 public class LogService {
 
     private final EventProducer<Log> producer;
-    private final AppService appService;
+    private final AppService_ASIS appService;
 
     public void saveLogs(String appKey, List<CreateLogRequest> requests) {
         Long appId = appService.getAppIdByToken(appKey);
