+++ src/main/kotlin/info/logbat/domain/project/application/AppService.kt
@@ -1,32 +1,13 @@
-package info.logbat.domain.project.application;
+package info.logbat.domain.project.application
 
-import info.logbat.domain.project.repository.AppRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.CacheConfig;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.stereotype.Service;
+import org.springframework.cache.annotation.CacheConfig
+import org.springframework.stereotype.Service
 
 @Service
-@RequiredArgsConstructor
-@CacheConfig(cacheNames = {"app"})
-public class AppService {
+@CacheConfig(cacheNames = ["app"])
+class AppService (
 
-    private static final String APP_NOT_FOUND_MESSAGE = "앱을 찾을 수 없습니다.";
-    public static final String APP_EXISTS_MESSAGE = "앱이 존재합니다.";
-    
-    private final AppRepository appRepository;
-
-    @Cacheable(key = "#token")
-    public Long getAppIdByToken(String token) {
-        return appRepository.getAppIdByToken(token)
-            .orElseThrow(() -> new IllegalArgumentException(APP_NOT_FOUND_MESSAGE));
-    }
-
-    @CacheEvict(key = "#token")
-    public void evictAppCache(String token) {
-        appRepository.getAppIdByToken(token).ifPresent(id -> {
-            throw new IllegalStateException(APP_EXISTS_MESSAGE);
-        });
-    }
+){
+    //private static final String APP_NOT_FOUND_MESSAGE = "앱을 찾을 수 없습니다.";
+    //    public static final String APP_EXISTS_MESSAGE = "앱이 존재합니다.";
 }
