+++ src/main/kotlin/info/logbat/domain/project/repository/AppRepository.kt
@@ -1,24 +1,14 @@
-package info.logbat.domain.project.repository;
+package info.logbat.domain.project.repository
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
+import org.springframework.jdbc.core.JdbcTemplate
+import org.springframework.stereotype.Repository
 
 @Repository
-@RequiredArgsConstructor
-public class AppRepository {
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public Optional<Long> getAppIdByToken(String token) {
-        String selectQuery = "SELECT id FROM apps WHERE app_key = UNHEX(REPLACE(?, '-', ''))";
-        try {
-            Long id = jdbcTemplate.queryForObject(selectQuery, Long.class, token);
-            return Optional.ofNullable(id);
-        } catch (Exception e) {
-            return Optional.empty();
-        }
+class AppRepository (
+    private val jdbcTemplate: JdbcTemplate
+){
+    fun getAppIdByToken(token: String): Long{
+        val selectQuery = "SELECT id FROM apps WHERE app_key = UNHEX(REPLACE(?, '-', ''))"
+        return jdbcTemplate.queryForObject(selectQuery, Long::class.java, token).or(0)
     }
 }
